requirements.txt
asgiref==3.2.3
cffi==1.13.2
Django==3.0.2
django-nose==1.4.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
nose==1.3.7
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
